qol: Reorganize files.
