+++ src/tests/matrix_trace.cpp
